% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptiveTableMunich2019dataset.R
\name{descriptiveTableMunich2019dataset}
\alias{descriptiveTableMunich2019dataset}
\title{descriptiveTableMunich2019dataset}
\usage{
descriptiveTableMunich2019dataset(savePath = NULL)
}
\arguments{
\item{savePath}{character string specifying the path and filename for exporting the table.\cr
If NULL (default), the table will not be exported.\cr
Expected format: "filepath/filename.extension", where extension can be svg, png, or pdf, the directory must exist beforehand.}
}
\value{
A 'table1' object containing descriptive statistics grouped by survival status.
}
\description{
Function to create, display and optionally export a table containing descriptive statistics for munich2019dataset.\cr
The table is createtd using the table1 package and is grouped by survival status.
}
\examples{
descriptiveTableMunich2019dataset()

descriptiveTableMunich2019dataset(savePath = "tables/munich_table.pdf")

}
