% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-global.R
\docType{data}
\name{global_variables}
\alias{global_variables}
\title{Declarations for glibal variables}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
global_variables
}
\description{
This file is used to suppress global variables NOTEs while running 'R CMD check',
 these variables are used as internal datasets in the package functions.
 They are declared in the 'data/' folder and are automatically loaded using
 'lazydata = TRUE'
}
\details{
@keywords internal
 @name global_variables
}
\keyword{datasets}
