% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveTable.R
\name{saveTable}
\alias{saveTable}
\title{saveTable}
\usage{
saveTable(t1, names, savePath = "")
}
\arguments{
\item{t1}{A 'table1' object to be saved.}

\item{names}{A character vector of at least 1 item: the dataset name,
other items may be used for testing}

\item{savePath}{A character string specifying the full path and filename for the exported table.\cr
The format must be: 'filepath/filename.extension' where:
\itemize{
  \item filepath is the directory (must already exist), default to the working directory;
  \item filename is the name of the file, default to default_filename;
  \item extension must be one of 'svg', 'png', or 'pdf', default is 'png'.
}}
}
\value{
A status number, 0 if the save was successfull, -1 if there were errors.
}
\description{
Internal function to save a table1 table object in svg, png or pdf format.
Default extension is png, default_filename is in this format: 'table_dataset_datetime.png'.
}
\keyword{internal}
