% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptiveTableUtrecht2019dataset.R
\name{descriptiveTableUtrecht2019dataset}
\alias{descriptiveTableUtrecht2019dataset}
\title{Descriptive statistic table for the utrecht2019dataset}
\usage{
descriptiveTableUtrecht2019dataset(savePath = NULL)
}
\arguments{
\item{savePath}{Character (optional). String specifying the path and filename for exporting the table.
\itemize{
\item If NULL (default), the table will not be exported.
\item If "" (empty string), the table will be saved in the working directory using a default name.
\item Expected format: 'filepath/filename.extensions' where:
\itemize{
\item filepath is the directory (must already exist), default is the working directory;
\item filename is the name of the file, default to 'table_utrecht2019dataset_timestamp.png';
\item extension must be one of png or pdf and must be specified, unless using the default filename.
}
}}
}
\value{
A 'table1' object containing descriptive statistics grouped by survival status.
}
\description{
Function to create, display and optionally export a table containing descriptive statistics for the utrecht2019dataset\cr
The table is created using the table1 package and is grouped by survival status.
}
\examples{
# Create and display the table, without saving anything
descriptiveTableUtrecht2019dataset()

## Not run:
# Create, display and save the table giving a path, filename and extension
descriptiveTableUtrecht2019dataset(savePath = "tables/utrecht_table.pdf")

# Create, display and save the table giving only the path, default name will be used
descriptiveTableUtrecht2019dataset(savePath = "tables/")

# Create, display and save the table giving a path and filename,default extension will be used
descriptiveTableUtrecht2019dataset(savePath = "tables/utrecht_table")

## End(Not run)

}
\seealso{
\code{\link[table1:table1]{table1::table1()}}
}
