% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glioblastomaEHRsData.R
\docType{package}
\name{glioblastomaEHRsData-package}
\alias{glioblastomaEHRsData}
\alias{glioblastomaEHRsData-package}
\title{glioblastomaEHRsData: Descriptive analysis on 3 EHRs datasets}
\description{
Il pacchetto fornisce dataset e funzioni per generare statistiche descrittive,
tabelle di frequenza, riepiloghi e grafici, facilitando l’analisi esplorativa.
}
\author{
\strong{Maintainer}: Samuele Marelli \email{samu2003.marelli@gmail.com}

}
\keyword{internal}
