% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tainan2020datasetDescriptiveStatistics.R
\name{Tainan2020datasetDescriptiveStatistics}
\alias{Tainan2020datasetDescriptiveStatistics}
\title{Tainan2020datasetDescriptiveStatistics}
\usage{
Tainan2020datasetDescriptiveStatistics(show = "all")
}
\arguments{
\item{show}{character string for the type of statistic to show
 \describe{
   \item{"continuous"}{Summary statistics for continuous variables.}
   \item{"categorical"}{Frequency tables for categorical variables.}
   \item{"all"}{Returns both summary statistics and frequency tables.}
}}
}
\value{
A list or an object depending on the value of \code{show}:
  \describe{
    \item{If \code{show = "continuous"}}{A data frame with descriptive statistics.}
    \item{If \code{show = "categorical"}}{A list of frequency tables.}
    \item{If \code{show = "all"}}{A list containing both descriptive statistics and frequency tables.}
  }
}
\description{
Provides descriptive statistics for the 'tainan2020dataset'
}
\examples{
  # Show both continuous and categorical statistics
  Tainan2020datasetDescriptiveStatistics(show = "all")

  # Only continuous variables
  Tainan2020datasetDescriptiveStatistics("continuous")

  # Only categorical variables
  Tainan2020datasetDescriptiveStatistics(show = "categorical")

}
